bugfixing, removed type casts

+++ Services/Mail/classes/Address/Type/class.ilMailAddressTypeHelper.php
@@ -54,7 +54,7 @@ class ilMailAddressTypeHelperImpl implements ilMailAddressTypeHelper
      */
     public function getUserIdByLogin(string $login) : int
     {
-        return (int) ilObjUser::getUserIdByLogin($login);
+        return ilObjUser::getUserIdByLogin($login);
     }
 
     /**
